Implemented webcam thread
I added a thread mechanism into the library so that when going into
streaming mode, the library creates a thread for the given webcam. In
this thread, the webcam will be grabbing images and storing it in its
own RGB buffer (.frame). There is a webcam_grab function that returns
the currently set RGB buffer.

The webcam instance also has a mtx_frame mutex to prevent reading the
frame while it is storing a new frame into place.